# android/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "picweight API",
    "description": "AI-assisted calorie & БЖУ tracker. Photograph what you eat; a server-side estimation agent returns calories and macros.",
    "license": {
      "name": "MIT"
    },
    "version": "master"
  },
  "paths": {
...
    "/api/v1/weights": {
      "get": {
        "tags": [
          "weights"
        ],
        "summary": "Weight history",
        "operationId": "list_weights",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive start date.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Inclusive end date.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum readings to return. Defaults to [`DEFAULT_WEIGHT_LIMIT`].",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Readings, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeightLogResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "weights"
        ],
        "summary": "Log a weight",
        "description": "Recomputes the user's targets from the new weight, since targets are derived from body data and must track it.",
        "operationId": "log_weight",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogWeightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The stored reading",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogWeightResponse"
                }
              }
            }
          },
          "400": {
            "description": "Implausible weight",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body returned for every [`AppError`].",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Generic for 5xx, specific for 4xx."
          }
        }
      },
...
      "LogWeightRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/weights`.",
        "required": [
          "weight_kg"
        ],
        "properties": {
          "logged_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When it was measured. Defaults to now."
          },
          "source": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WeightSource",
                "description": "How it was captured. Defaults to `manual`."
              }
            ]
          },
          "weight_kg": {
            "type": "number",
            "format": "double",
            "description": "Weight in kilograms."
          }
        }
      },
      "LogWeightResponse": {
        "type": "object",
        "description": "Result of logging a weight.",
        "required": [
          "weight",
          "targets_recomputed"
        ],
        "properties": {
          "targets_recomputed": {
            "type": "boolean",
            "description": "True when this reading caused targets to be recomputed."
          },
          "weight": {
            "$ref": "#/components/schemas/WeightLogResponse",
            "description": "The stored reading."
          }
        }
      },
...
      "WeightLogResponse": {
        "type": "object",
        "description": "One weight reading.",
        "required": [
          "id",
          "logged_at",
          "weight_kg",
          "source"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "`weight_logs.id`."
          },
          "logged_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it was measured."
          },
          "source": {
            "$ref": "#/components/schemas/WeightSource",
            "description": "How it was captured."
          },
          "weight_kg": {
            "type": "number",
            "format": "double",
            "description": "Weight in kilograms."
          }
        }
      },
      "WeightSource": {
        "type": "string",
        "description": "How a weight reading was captured.",
        "enum": [
          "Manual",
          "Onboarding",
          "Import",
          "Scale"
        ]
      }
    },
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session JWT. Web clients receive it as the HttpOnly `picweight_session` cookie after OIDC login; native clients obtain it from `POST /api/auth/token` and send it as `Authorization: Bearer`."
      }
    }
  },
...
}